Category and Contact create, edit, and many to many relationship works.

+++ src/ContactPro/Controllers/CategoriesController.cs
@@ -1,20 +1,21 @@
-
-using System.Threading;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using JHipsterNet.Core.Pagination;
-using ContactPro.Domain.Entities;
 using ContactPro.Crosscutting.Exceptions;
+using ContactPro.Domain.Entities;
+using ContactPro.Domain.Repositories.Interfaces;
+using ContactPro.Domain.Services;
+using ContactPro.Infrastructure.Data;
+using ContactPro.Infrastructure.Web.Rest.Utilities;
 using ContactPro.Web.Extensions;
 using ContactPro.Web.Filters;
 using ContactPro.Web.Rest.Utilities;
-using ContactPro.Domain.Repositories.Interfaces;
-using ContactPro.Domain.Services.Interfaces;
-using ContactPro.Infrastructure.Web.Rest.Utilities;
+using JHipsterNet.Core.Pagination;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Microsoft.EntityFrameworkCore;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace ContactPro.Controllers
 {
@@ -26,12 +27,20 @@ namespace ContactPro.Controllers
         private const string EntityName = "category";
         private readonly ILogger<CategoriesController> _log;
         private readonly ICategoryRepository _categoryRepository;
+        private readonly ApplicationDatabaseContext _context;
+        private UtilityService _utilityService;
+
 
         public CategoriesController(ILogger<CategoriesController> log,
-        ICategoryRepository categoryRepository)
+        ICategoryRepository categoryRepository,
+        ApplicationDatabaseContext context,
+        UserManager<User> userManager,
+        IHttpContextAccessor httpContextAccessor)
         {
             _log = log;
             _categoryRepository = categoryRepository;
+            _context = context;
+            _utilityService = new UtilityService(userManager, httpContextAccessor);
         }
 
         [HttpPost]
@@ -42,6 +51,10 @@ namespace ContactPro.Controllers
             if (category.Id != 0)
                 throw new BadRequestAlertException("A new category cannot already have an ID", EntityName, "idexists");
 
+            category.Created = _utilityService.GetNowInUtc();           
+            category.UserId = _utilityService.GetCurrentUserId();
+            await CheckAllContacts(category);
+            
             await _categoryRepository.CreateOrUpdateAsync(category);
             await _categoryRepository.SaveChangesAsync();
             return CreatedAtAction(nameof(GetCategory), new { id = category.Id }, category)
@@ -55,6 +68,12 @@ namespace ContactPro.Controllers
             _log.LogDebug($"REST request to update Category : {category}");
             if (category.Id == 0) throw new BadRequestAlertException("Invalid Id", EntityName, "idnull");
             if (id != category.Id) throw new BadRequestAlertException("Invalid Id", EntityName, "idinvalid");
+            if (!_utilityService.GetCurrentUserId().Equals(category.UserId)) throw new BadRequestAlertException("User Id and Category User Id doesn't match", EntityName, "useridmatch");            
+            Category oldCategory = await _categoryRepository.QueryHelper().GetOneAsync(c => c.Id == id && _utilityService.GetCurrentUserId().Equals(c.UserId));
+            if (oldCategory == null) throw new BadRequestAlertException("Invalid Category Id or User Id", EntityName, "idnull");
+            
+            await CheckAllContacts(category);
+
             await _categoryRepository.CreateOrUpdateAsync(category);
             await _categoryRepository.SaveChangesAsync();
             return Ok(category)
@@ -66,8 +85,7 @@ namespace ContactPro.Controllers
         {
             _log.LogDebug("REST request to get a page of Categories");
             var result = await _categoryRepository.QueryHelper()
-                .Include(category => category.User)
-                .Include(category => category.Contacts)
+                .Filter(c => c.UserId != null && _utilityService.GetCurrentUserId().Equals(c.UserId))                
                 .GetPageAsync(pageable);
             return Ok(result.Content).WithHeaders(result.GeneratePaginationHttpHeaders());
         }
@@ -77,9 +95,8 @@ namespace ContactPro.Controllers
         {
             _log.LogDebug($"REST request to get Category : {id}");
             var result = await _categoryRepository.QueryHelper()
-                .Include(category => category.User)
                 .Include(category => category.Contacts)
-                .GetOneAsync(category => category.Id == id);
+                .GetOneAsync(category => category.Id == id && _utilityService.GetCurrentUserId().Equals(category.UserId));
             return ActionResultUtil.WrapOrNotFound(result);
         }
 
@@ -87,9 +104,23 @@ namespace ContactPro.Controllers
         public async Task<IActionResult> DeleteCategory([FromRoute] long id)
         {
             _log.LogDebug($"REST request to delete Category : {id}");
+
+            Category category = await _categoryRepository.QueryHelper()
+                .GetOneAsync(category => category.Id == id && _utilityService.GetCurrentUserId().Equals(category.UserId));
+            if (category == null) throw new BadRequestAlertException("Invalid Id", EntityName, "idnull");
+
             await _categoryRepository.DeleteByIdAsync(id);
             await _categoryRepository.SaveChangesAsync();
             return NoContent().WithHeaders(HeaderUtil.CreateEntityDeletionAlert(EntityName, id.ToString()));
+        }
+
+        private async Task CheckAllContacts(Category category)
+        {
+            foreach(Contact contact in category.Contacts)
+            {
+                Contact fetchedContact = await _context.Contacts.FirstOrDefaultAsync(c => c.Id == contact.Id && _utilityService.GetCurrentUserId().Equals(c.UserId));
+                if (fetchedContact is null) throw new BadRequestAlertException("Invalid Contact Id", EntityName, "idinvalid");
+            }
         }
     }
 }
